Make Off-line Collection More Tolerant
It is not uncommon for offline data collection to be aborted, or on
devices that do not have such checks never started.  This removes
the logic that bombs out returning WARNING or UNKNOWN and just echos
the status to the output.  Also corrects a formatting problem which
didn't take into account the longest output text