Move to C++, Log Checks & Performance Data
Fairly large refactor to begin collecting SMART logs from the device as well as
standard attribute metrics, which may and do hint above and beyond those
already collected that failure is impending.  In the process of deprecating
verbose output in preference of performance data that may be graphed and have
individual upper and lower thresholds applied eventually it was deemed in the
best interests to port the check to C++ for cleanliness of handling buffering,
formatting and outputting of metrics.