Extend SmartID Class to SmartAttribute
Updates the SmartID wrapper to encapsulate a whole SMART attribute which
has the knock on effect of encapsulating the coversion from a split raw
value to a 48bit scalar, and also the output formatting for specific IDs.
This also updates unsafe C style casts into C++ static and reinterpret
casts where appropriate.