Support PERC and CCISS Controllers
The error checking in release 1.2.1 had the unintended consequence of revealing
that HP and Dell RAID controllers don't support ATA tunnelling.  This change
removes the verbose debugging and instead propagates the error condition up the
call stack.  If the initial ATA IDENTITY call fails then we determine that ATA
tunneling is not supported by this particular device and return an OK status.